[UPD] Rewrited OMSUser with new config
